refactor: Split training functions into separate modules.
