+++ src/training/train_model.py
@@ -1,6 +1,6 @@
 import torch
-import torch.nn as nn
-import torch.optim as optim
+# import torch.nn as nn
+# import torch.optim as optim
 import torch.optim.lr_scheduler
 from torch.utils.data import DataLoader
 from torch.utils.data import Dataset
@@ -8,76 +8,10 @@ from torch.utils.data import Dataset
 from model_definitions.parabolic_motion_model import ParabolicMotionModel
 from utils.config_utils import load_config
 from utils.validation_utils import verify_motion_against_params
-from utils.data_utils import batch_shuffle
-
-def initialize_model(input_size: int, hidden_size: int, output_size: int) -> nn.Module:
-    """Initialize the ParabolicMotionModel with the given sizes."""
-    model = ParabolicMotionModel(input_size, hidden_size, output_size)
-    return model
-
-def configure_training(learning_rate: float, model: nn.Module) -> tuple:
-    """Set up the loss function and optimizer."""
-    criterion = nn.MSELoss()  # Evaluate loss by MSE.
-    optimizer = optim.Adam(model.parameters(), lr=learning_rate)  # Optimize model weights by Adam.
-    return criterion, optimizer
-
-def train_one_epoch(model: nn.Module, data_loader: DataLoader, criterion: nn.Module, optimizer: optim.Optimizer, epoch: int, epoch_max: int) -> float:
-    """Train the model for one epoch."""
-    for i, (motion, params) in enumerate(data_loader):
-        # Fetch the first elements of the parameters tensor, because they are the same for all time steps
-        unique_params = params[0].unsqueeze(0)
-
-        # Fetch the model outputs
-        outputs = model(unique_params)
-
-        # Calculate the target values, i.e. the reaching distance and max height
-        target_reaching_distance = motion[-1, 1]
-        target_max_height = motion[:, 2].max()
-
-        # Create target tensor
-        target = torch.tensor([target_reaching_distance, target_max_height]).unsqueeze(0)
-
-        # Calculate the loss
-        loss = criterion(outputs, target)
-
-        # Reset gradients
-        optimizer.zero_grad()
-
-        # Backpropagate the loss and optimize the weights
-        loss.backward()
-        optimizer.step()
-
-    return loss.item()
-
-def save_model(model, file_path):
-    """Save the PyTorch model to the specified file path."""
-    torch.save(model.state_dict(), file_path)
-
-def validate_model(model: nn.Module, val_loader: DataLoader, criterion: nn.Module) -> float:
-    """Validate the model using the validation dataset."""
-    model.eval()  # Set the model to evaluation mode
-    val_loss = 0.0
-
-    with torch.no_grad():
-        for i, (motion, params) in enumerate(val_loader):
-            # Fetch the first elements of the parameters tensor, because they are the same for all time steps
-            unique_params = params[0].unsqueeze(0)
-
-            # Fetch the model outputs from input parameters
-            outputs = model(unique_params)
-
-            # Calculate the target values, i.e. the reaching distance and max height
-            target_reaching_distance = motion[-1, 1]
-            target_max_height = motion[:, 2].max()
-
-            # Create target tensor
-            target = torch.tensor([target_reaching_distance, target_max_height]).unsqueeze(0)
-
-            # Calculate the loss
-            loss = criterion(outputs, target)
-            val_loss += loss.item()
-
-    return val_loss / len(val_loader)
+from utils.data_utils import batch_shuffle, save_model
+from training.epoch_train import train_one_epoch
+from training.inference_model_validation import validate_model
+from training.training_setup import configure_training
 
 def train_model(train_dataset: Dataset, val_dataset: Dataset, batch_size: int) -> None:
     """Main training loop."""
@@ -90,9 +24,9 @@ def train_model(train_dataset: Dataset, val_dataset: Dataset, batch_size: int) -
     hidden_size = int(cfg['training']['hidden_size'])
     
     # Initialize the model
-    model = initialize_model(input_size=2,
-                             hidden_size=hidden_size,
-                             output_size=2)
+    model = ParabolicMotionModel(input_size=2,
+                                 hidden_size=hidden_size,
+                                 output_size=2)
 
     # Set the loss function and optimizer
     criterion, optimizer = configure_training(init_learning_rate, model)
