+++ src/utils/data_utils.py
@@ -12,3 +12,7 @@ def batch_shuffle(dataset: Dataset, batch_size: int) -> torch.Tensor:
     # Shuffle the order of batches
     indices = indices[torch.randperm(n_batches)]
     return indices.view(-1)
+
+def save_model(model, file_path):
+    """Save the PyTorch model to the specified file path."""
+    torch.save(model.state_dict(), file_path)
